[AOT Refactor] utils
[ghstack-poisoned]
